moved rendered from App.js over to PostsPage component
